+++ app/src/main/java/com/teamx/hatlyUser/dataclasses/Product.kt
@@ -6,7 +6,7 @@ import androidx.annotation.Keep
 data class Product(
     val order_quantity: Int,
     val product_id: String,
-    val subtotal: Int,
-    val unit_price: Int,
+    val subtotal: Double,
+    val unit_price: Double,
     val variation_option_id: String
 )
