+++ app/src/main/java/com/teamx/hatlyUser/ui/fragments/payments/cart/model/ModelCart.kt
@@ -5,7 +5,7 @@ import androidx.annotation.Keep
 @Keep
 data class ModelCart(
     val products: List<Product>?,
-    val subTotal: Int,
+    val subTotal: Double,
     val tax: Double,
     val total: Double
 )
