+++ app/src/main/java/com/teamx/hatlyUser/ui/fragments/profile/orderhistory/model/Doc.kt
@@ -18,7 +18,7 @@ data class Doc(
     val shop: Shop,
     val specialNote: String,
     var status: String,
-    val subTotal: Int,
+    val subTotal: Double,
     val tax: Double,
     val total: Double,
     val useWallet: Boolean,
